+++ Puzzle/cpuclass.h
@@ -33,9 +33,9 @@ public:
 	int GetCpuPercentage();
 
 private:
-	bool			m_canReadCpu;
-	HQUERY			m_queryHandle;
-	HCOUNTER		m_counterHandle;
-	unsigned long	m_lastSampleTime;
-	long			m_cpuUsage;
+	bool				m_canReadCpu;
+	HQUERY				m_queryHandle;
+	HCOUNTER			m_counterHandle;
+	unsigned long long	m_lastSampleTime;
+	long				m_cpuUsage;
 };
